corrected finance calculator code

+++ python_introduction/finance_calculator.py
@@ -1,6 +1,6 @@
 # Prompt the user for their monthly income and monthly expenses
-monthly_income = int(input("Enter your monthly income: "))
-monthly_expenses = int(input("Enter your total monthly expenses: "))
+monthly_income = float(input("Enter your monthly income: "))
+monthly_expenses = float(input("Enter your total monthly expenses: "))
 
 # Calculating the monthly savings
 monthly_savings = monthly_income - monthly_expenses
